make pomodoro hook
