--- pho/src/config/pomodoro.json
@@ -1,25 +1,0 @@
-{
-  "pomodoroes": [
-    {
-      "sessionInterval": 2,
-      "breakInterval": 1,
-      "sessionName": "On Pomodoro 1"
-    },
-    {
-      "sessionInterval": 3,
-      "breakInterval": 1,
-      "sessionName": "On Pomodoro 2"
-    },
-    {
-      "sessionInterval": 5,
-      "breakInterval": 2,
-      "sessionName": "On Pomodoro 3"
-    },
-    {
-      "sessionInterval": 8,
-      "breakInterval": 4,
-      "sessionName": "On Pomodoro 4"
-    }
-  ],
-  "longBreak": 8
-}
